requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
google-genai==1.3.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
pgvector==0.3.6
psycopg2-binary==2.9.10
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
numpy==2.2.1
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1
